+++ test/flake-parts/flake.nix
@@ -101,6 +101,10 @@
                         cat ${escapeShellArg cfg.configLocal}"/settings_user.yml" \
                           | grep -F ${escapeShellArg (inputs.conan-flake.lib.versionFromStdenvCc lib llvmPackages.libcxxStdenv)}
 
+                        cat ${escapeShellArg cfg.configLocal}"/global.conf" \
+                          | grep -F "core.graph:compatibility_mode" \
+                          | grep -F "optimized"
+
                         cat ${escapeShellArg cfg.configLocal}"/profiles/default" \
                           | grep -F "build_type="${escapeShellArg cfg.final.profiles.settings._.build_type}
                         cat ${escapeShellArg cfg.configLocal}"/profiles/default" \
@@ -257,6 +261,10 @@
                   | grep -F ${escapeShellArg (inputs.conan-flake.lib.versionFromStdenvCc lib backendStdenv)}
                 cat "${configuration}/config/settings_user.yml" \
                   | grep -F ${escapeShellArg (inputs.conan-flake.lib.versionFromStdenvCc lib llvmPackages.libcxxStdenv)}
+
+                cat "${configuration}/config/global.conf" \
+                  | grep -F "core.graph:compatibility_mode" \
+                  | grep -F "optimized"
 
                 cat "${configuration}/config/profiles/default" \
                   | grep -F "build_type="${escapeShellArg profiles.settings._.build_type}